+++ anyblok_marshmallow/tests/test_fields.py
@@ -7,11 +7,13 @@
 # obtain one at http://mozilla.org/MPL/2.0/.
 from anyblok.tests.testcase import DBTestCase
 from anyblok_marshmallow import ModelSchema
-from anyblok_marshmallow.fields import Nested
+from anyblok_marshmallow.fields import Nested, File
 from anyblok import Declarations
-from anyblok.column import Integer
+from anyblok.column import Integer, LargeBinary
 from anyblok.field import Function
 from anyblok.relationship import One2One
+from os import urandom
+from base64 import b64encode
 
 
 class TestField(DBTestCase):
@@ -36,6 +38,13 @@ class TestField(DBTestCase):
         class Exemple2:
             id = Integer(primary_key=True)
             exemple = One2One(model='Model.Exemple', backref='exemple2')
+
+    def add_field_largebinary(self):
+
+        @Declarations.register(Declarations.Model)
+        class Exemple:
+            id = Integer(primary_key=True)
+            file = LargeBinary()
 
     def test_dump_function(self):
         registry = self.init_registry(self.add_field_function)
@@ -107,6 +116,17 @@ class TestField(DBTestCase):
                 model = 'Model.Exemple2'
 
         return Exemple2Schema
+
+    def getExempleSchemaLO(self):
+
+        class ExempleSchema(ModelSchema):
+
+            file = File()
+
+            class Meta:
+                model = 'Model.Exemple'
+
+        return ExempleSchema
 
     def test_dump_one2one_1(self):
         registry = self.init_registry(self.add_field_one2one)
@@ -198,4 +218,42 @@ class TestField(DBTestCase):
         }
         exemple2_schema = self.getExemple2Schema()(registry=registry)
         errors = exemple2_schema.validate(dump_data)
+        self.assertFalse(errors)
+
+    def test_dump_file(self):
+        registry = self.init_registry(self.add_field_largebinary)
+        exemple_schema = self.getExempleSchemaLO()(registry=registry)
+        file_ = urandom(100)
+        exemple = registry.Exemple.insert(file=file_)
+        data, errors = exemple_schema.dump(exemple)
+        self.assertFalse(errors)
+        self.assertEqual(
+            data,
+            {
+                'id': exemple.id,
+                'file': b64encode(file_).decode('utf-8'),
+            }
+        )
+
+    def test_load_file(self):
+        registry = self.init_registry(self.add_field_function)
+        file_ = urandom(100)
+        dump_data = {
+            'id': 1,
+            'file': b64encode(file_).decode('utf-8'),
+        }
+        exemple_schema = self.getExempleSchemaLO()(registry=registry)
+        data, errors = exemple_schema.load(dump_data)
+        self.assertEqual(data, {'id': 1, 'file': file_})
+        self.assertFalse(errors)
+
+    def test_validate_file(self):
+        registry = self.init_registry(self.add_field_function)
+        file_ = urandom(100)
+        dump_data = {
+            'id': 1,
+            'file': b64encode(file_).decode('utf-8'),
+        }
+        exemple_schema = self.getExempleSchemaLO()(registry=registry)
+        errors = exemple_schema.validate(dump_data)
         self.assertFalse(errors)